fix(api): rename test utils file from utils.py to helpers.py
- Rename api/tests/utils.py to api/tests/helpers.py for better naming convention
- Update all import statements in test files to use new helpers module
- Maintain all existing functionality with improved file organization

This change improves code organization by using a more descriptive filename
for test helper functions.
